pyproject.toml
[project]
name = "timedesign-backend"
version = "0.1.0"
description = "timedesign-backend application"
authors = [{ name = "Shota Onishi" }]
dependencies = [
    "aws-cdk-lib>=2.135.0",
    "aws-cdk.aws-lambda-python-alpha>=2.136.0a0",
    "aws-lambda-powertools>=2.36.0",
    "constructs>=10.0.0,<11.0.0",
    "pydantic>=2.6.4",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = ["pytest>=8.1.1"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/time_design_backend"]

[tool.ruff]
line-length = 80
indent-width = 4
select = [
    "F", # pyflakes
    "E", # pycodestyle
    "W", # pycodestyle warnings
    "I", # isort
    "D", # pydocstyle
]
extend-ignore = [
    "D100", # undocumented-public-module
    "D101", # undocumented-public-class
    "D105", # undocumented-magic-method
    "D107", # undocumented-public-init
    "D205", # blank-line-after-summary
    "D415", # ends-in-punctuation
    # "I001", # 
]

[tool.ruff.lint.pydocstyle]
convention = "google" # docstringをgoogle styleにする

[tool.mypy]
strict = true
ignore_missing_imports = true
disallow_untyped_decorators = false
disallow_untyped_calls = false
